+++ soane/comms/__init__.py
@@ -8,3 +8,4 @@ from soane.comms.delete import delete
 from soane.comms.list   import list_
 from soane.comms.open   import open_
 from soane.comms.read   import read
+from soane.comms.rename import rename